videocontrol: move player to separate file
